Consolidate CLI E2E smoke tests
