--- tests/Aspire.Cli.EndToEnd.Tests/BundleSmokeTests.cs
@@ -1,47 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using Aspire.Cli.EndToEnd.Tests.Helpers;
-using Aspire.Cli.Tests.Utils;
-using Hex1b.Automation;
-using Xunit;
-
-namespace Aspire.Cli.EndToEnd.Tests;
-
-/// <summary>
-/// End-to-end tests for running .NET csproj AppHost projects using the Aspire bundle.
-/// Validates that the bundle correctly provides DCP and Dashboard paths to the hosting
-/// infrastructure when running SDK-based app hosts (not just polyglot/guest app hosts).
-/// </summary>
-public sealed class BundleSmokeTests(ITestOutputHelper output)
-{
-    [CaptureWorkspaceOnFailure]
-    [Fact]
-    public async Task CreateAndRunAspireStarterProjectWithBundle()
-    {
-        var repoRoot = CliE2ETestHelpers.GetRepoRoot();
-        var strategy = CliInstallStrategy.Detect();
-
-        var workspace = TemporaryWorkspace.Create(output);
-
-        using var terminal = CliE2ETestHelpers.CreateDockerTestTerminal(repoRoot, strategy, output, mountDockerSocket: true, workspace: workspace);
-
-        var pendingRun = terminal.RunAsync(TestContext.Current.CancellationToken);
-
-        var counter = new SequenceCounter();
-        var auto = new Hex1bTerminalAutomator(terminal, defaultTimeout: TimeSpan.FromSeconds(500));
-
-        await auto.PrepareDockerEnvironmentAsync(counter, workspace);
-        await auto.InstallAspireCliAsync(strategy, counter);
-
-        await auto.AspireNewAsync("BundleStarterApp", counter);
-
-        await auto.AspireStartAsync(counter);
-        await auto.AspireStopAsync(counter);
-
-        await auto.TypeAsync("exit");
-        await auto.EnterAsync();
-
-        await pendingRun;
-    }
-}
