--- tests/Aspire.Cli.EndToEnd.Tests/EmptyAppHostTemplateTests.cs
@@ -1,50 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using Aspire.Cli.EndToEnd.Tests.Helpers;
-using Aspire.Cli.Tests.Utils;
-using Hex1b.Automation;
-using Xunit;
-
-namespace Aspire.Cli.EndToEnd.Tests;
-
-/// <summary>
-/// End-to-end tests for Aspire CLI with Empty AppHost template.
-/// Each test class runs as a separate CI job for parallelization.
-/// </summary>
-public sealed class EmptyAppHostTemplateTests(ITestOutputHelper output)
-{
-    [CaptureWorkspaceOnFailure]
-    [Fact]
-    public async Task CreateAndRunEmptyAppHostProject()
-    {
-        var repoRoot = CliE2ETestHelpers.GetRepoRoot();
-        var strategy = CliInstallStrategy.Detect();
-        var workspace = TemporaryWorkspace.Create(output);
-
-        using var terminal = CliE2ETestHelpers.CreateDockerTestTerminal(repoRoot, strategy, output, mountDockerSocket: true, workspace: workspace);
-
-        var pendingRun = terminal.RunAsync(TestContext.Current.CancellationToken);
-
-        var counter = new SequenceCounter();
-        var auto = new Hex1bTerminalAutomator(terminal, defaultTimeout: TimeSpan.FromSeconds(500));
-
-        await auto.PrepareDockerEnvironmentAsync(counter, workspace);
-        await auto.InstallAspireCliAsync(strategy, counter);
-
-        await auto.AspireNewAsync("AspireEmptyApp", counter, template: AspireTemplate.EmptyAppHost);
-
-        // Start the empty AppHost to verify the scaffolded project works
-        await auto.TypeAsync("cd AspireEmptyApp");
-        await auto.EnterAsync();
-        await auto.WaitForSuccessPromptAsync(counter);
-
-        await auto.AspireStartAsync(counter);
-        await auto.AspireStopAsync(counter);
-
-        await auto.TypeAsync("exit");
-        await auto.EnterAsync();
-
-        await pendingRun;
-    }
-}
