+++ tests/Aspire.Cli.EndToEnd.Tests/SmokeTests.cs
@@ -10,8 +10,7 @@ using Xunit;
 namespace Aspire.Cli.EndToEnd.Tests;
 
 /// <summary>
-/// End-to-end tests for Aspire CLI run command (creating and launching projects).
-/// Each test class runs as a separate CI job for parallelization.
+/// End-to-end smoke tests for the core Aspire CLI template scenarios.
 /// </summary>
 public sealed class SmokeTests(ITestOutputHelper output)
 {
@@ -38,7 +37,7 @@ public sealed class SmokeTests(ITestOutputHelper output)
         await auto.InstallAspireCliAsync(strategy, counter);
 
         // Create a new project using aspire new
-        await auto.AspireNewAsync("AspireStarterApp", counter);
+        await auto.AspireNewAsync("AspireStarterApp", counter, useRedisCache: false);
 
         // Run the project with aspire run
         await auto.TypeAsync("aspire run");
@@ -55,14 +54,233 @@ public sealed class SmokeTests(ITestOutputHelper output)
                     "Unexpected apphost selection prompt detected! " +
                     "This indicates multiple apphosts were incorrectly detected.");
             }
-            return s.ContainsText("Press CTRL+C to stop the AppHost and exit.");
-        }, timeout: TimeSpan.FromMinutes(2), description: "Press CTRL+C message (aspire run started)");
+            return s.ContainsText("Press CTRL+C to stop the AppHost and exit.")
+                || s.ContainsText("Press CTRL+C to stop the apphost and exit.");
+        }, timeout: TimeSpan.FromMinutes(5), description: "Press CTRL+C message (aspire run started)");
 
         // Stop the running apphost with Ctrl+C
         await auto.Ctrl().KeyAsync(Hex1bKey.C);
         await auto.WaitForSuccessPromptAsync(counter);
 
         // Exit the shell
+        await auto.TypeAsync("exit");
+        await auto.EnterAsync();
+
+        await pendingRun;
+    }
+
+    /// <summary>
+    /// Creates a starter project, starts it with aspire start, and verifies the
+    /// starter resources come up and the web frontend endpoint responds with HTTP 200.
+    /// Catches runtime regressions where templates build but fail to serve traffic.
+    /// </summary>
+    [Fact]
+    public async Task StarterTemplateEndpointsRespond()
+    {
+        var repoRoot = CliE2ETestHelpers.GetRepoRoot();
+        var strategy = CliInstallStrategy.Detect();
+        var workspace = TemporaryWorkspace.Create(output);
+
+        using var terminal = CliE2ETestHelpers.CreateDockerTestTerminal(repoRoot, strategy, output, mountDockerSocket: true, workspace: workspace);
+
+        var pendingRun = terminal.RunAsync(TestContext.Current.CancellationToken);
+
+        var counter = new SequenceCounter();
+        var auto = new Hex1bTerminalAutomator(terminal, defaultTimeout: TimeSpan.FromSeconds(500));
+
+        await auto.PrepareDockerEnvironmentAsync(counter, workspace);
+        await auto.InstallAspireCliAsync(strategy, counter);
+
+        await auto.AspireNewAsync("EndpointTest", counter, useRedisCache: false);
+
+        await auto.TypeAsync("cd EndpointTest");
+        await auto.EnterAsync();
+        await auto.WaitForSuccessPromptAsync(counter);
+
+        await auto.AspireStartAsync(counter);
+
+        await auto.TypeAsync("aspire wait webfrontend --status up --timeout 300");
+        await auto.EnterAsync();
+        await auto.WaitUntilTextAsync("is up (running).", timeout: TimeSpan.FromMinutes(6));
+        await auto.WaitForSuccessPromptAsync(counter);
+
+        await auto.AssertResourcesExistAsync(counter, "webfrontend", "apiservice");
+
+        await auto.TypeAsync("aspire describe webfrontend --format json > webfrontend.json");
+        await auto.EnterAsync();
+        await auto.WaitForSuccessPromptAsync(counter);
+
+        await auto.TypeAsync("WEB_URL=$(grep -oE 'https?://localhost:[0-9]+' webfrontend.json | head -1); echo \"$WEB_URL\"");
+        await auto.EnterAsync();
+        await auto.WaitUntilAsync(
+            s => s.ContainsText("http://localhost:") || s.ContainsText("https://localhost:"),
+            timeout: TimeSpan.FromSeconds(30),
+            description: "web frontend URL");
+        await auto.WaitForSuccessPromptAsync(counter);
+
+        await auto.TypeAsync("curl -ksSL -o /dev/null -w 'webfrontend-http-%{http_code}' \"$WEB_URL\" || echo 'webfrontend-http-failed'");
+        await auto.EnterAsync();
+        await auto.WaitUntilTextAsync("webfrontend-http-200", timeout: TimeSpan.FromSeconds(30));
+        await auto.WaitForSuccessPromptAsync(counter);
+
+        await auto.AspireStopAsync(counter);
+
+        await auto.TypeAsync("exit");
+        await auto.EnterAsync();
+
+        await pendingRun;
+    }
+
+    [CaptureWorkspaceOnFailure]
+    [Fact]
+    public async Task CreateAndRunEmptyAppHostProject()
+    {
+        var repoRoot = CliE2ETestHelpers.GetRepoRoot();
+        var strategy = CliInstallStrategy.Detect();
+        var workspace = TemporaryWorkspace.Create(output);
+
+        using var terminal = CliE2ETestHelpers.CreateDockerTestTerminal(repoRoot, strategy, output, mountDockerSocket: true, workspace: workspace);
+
+        var pendingRun = terminal.RunAsync(TestContext.Current.CancellationToken);
+
+        var counter = new SequenceCounter();
+        var auto = new Hex1bTerminalAutomator(terminal, defaultTimeout: TimeSpan.FromSeconds(500));
+
+        await auto.PrepareDockerEnvironmentAsync(counter, workspace);
+        await auto.InstallAspireCliAsync(strategy, counter);
+
+        await auto.AspireNewAsync("AspireEmptyApp", counter, template: AspireTemplate.EmptyAppHost);
+
+        await auto.TypeAsync("cd AspireEmptyApp");
+        await auto.EnterAsync();
+        await auto.WaitForSuccessPromptAsync(counter);
+
+        await auto.AspireStartAsync(counter);
+        await auto.AspireStopAsync(counter);
+
+        await auto.TypeAsync("exit");
+        await auto.EnterAsync();
+
+        await pendingRun;
+    }
+
+    [CaptureWorkspaceOnFailure]
+    [Fact]
+    public async Task CreateAndRunTypeScriptStarterProject()
+    {
+        var repoRoot = CliE2ETestHelpers.GetRepoRoot();
+        var strategy = CliInstallStrategy.Detect();
+        var workspace = TemporaryWorkspace.Create(output);
+
+        using var terminal = CliE2ETestHelpers.CreateDockerTestTerminal(repoRoot, strategy, output, mountDockerSocket: true, workspace: workspace);
+
+        var pendingRun = terminal.RunAsync(TestContext.Current.CancellationToken);
+
+        var counter = new SequenceCounter();
+        var auto = new Hex1bTerminalAutomator(terminal, defaultTimeout: TimeSpan.FromSeconds(500));
+
+        await auto.PrepareDockerEnvironmentAsync(counter, workspace);
+        await auto.InstallAspireCliAsync(strategy, counter);
+
+        await auto.AspireNewAsync("TsStarterApp", counter, template: AspireTemplate.ExpressReact);
+
+        var projectRoot = Path.Combine(workspace.WorkspaceRoot.FullName, "TsStarterApp");
+        var modulesDir = Path.Combine(projectRoot, ".modules");
+
+        if (!Directory.Exists(modulesDir))
+        {
+            throw new InvalidOperationException($".modules directory was not created at {modulesDir}");
+        }
+
+        var aspireModulePath = Path.Combine(modulesDir, "aspire.ts");
+        if (!File.Exists(aspireModulePath))
+        {
+            throw new InvalidOperationException($"Expected generated file not found: {aspireModulePath}");
+        }
+
+        await auto.TypeAsync("cd TsStarterApp");
+        await auto.EnterAsync();
+        await auto.WaitForSuccessPromptAsync(counter);
+
+        await auto.AspireStartAsync(counter);
+        await auto.AspireStopAsync(counter);
+
+        await auto.TypeAsync("exit");
+        await auto.EnterAsync();
+
+        await pendingRun;
+    }
+
+    [Fact]
+    [CaptureWorkspaceOnFailure]
+    public async Task CreateAndRunPythonReactProject()
+    {
+        var repoRoot = CliE2ETestHelpers.GetRepoRoot();
+        var strategy = CliInstallStrategy.Detect();
+        var workspace = TemporaryWorkspace.Create(output);
+
+        using var terminal = CliE2ETestHelpers.CreateDockerTestTerminal(repoRoot, strategy, output, mountDockerSocket: true, workspace: workspace);
+
+        var pendingRun = terminal.RunAsync(TestContext.Current.CancellationToken);
+
+        var counter = new SequenceCounter();
+        var auto = new Hex1bTerminalAutomator(terminal, defaultTimeout: TimeSpan.FromSeconds(500));
+
+        await auto.PrepareDockerEnvironmentAsync(counter, workspace);
+        await auto.InstallAspireCliAsync(strategy, counter);
+
+        await auto.AspireNewAsync("AspirePyReactApp", counter, template: AspireTemplate.PythonReact, useRedisCache: false);
+
+        await auto.TypeAsync("cd AspirePyReactApp");
+        await auto.EnterAsync();
+        await auto.WaitForSuccessPromptAsync(counter);
+
+        await auto.AspireStartAsync(counter);
+        await auto.AspireStopAsync(counter);
+
+        await auto.TypeAsync("exit");
+        await auto.EnterAsync();
+
+        await pendingRun;
+    }
+
+    [CaptureWorkspaceOnFailure]
+    [Fact]
+    public async Task CreateAndRunAspireStarterProjectWithBundle()
+    {
+        var repoRoot = CliE2ETestHelpers.GetRepoRoot();
+        var strategy = CliInstallStrategy.Detect();
+        var workspace = TemporaryWorkspace.Create(output);
+
+        using var terminal = CliE2ETestHelpers.CreateDockerTestTerminal(repoRoot, strategy, output, mountDockerSocket: true, workspace: workspace);
+
+        var pendingRun = terminal.RunAsync(TestContext.Current.CancellationToken);
+
+        var counter = new SequenceCounter();
+        var auto = new Hex1bTerminalAutomator(terminal, defaultTimeout: TimeSpan.FromSeconds(500));
+
+        await auto.PrepareDockerEnvironmentAsync(counter, workspace);
+        await auto.InstallAspireCliAsync(strategy, counter);
+
+        await auto.AspireNewAsync("BundleStarterApp", counter, useRedisCache: false);
+
+        await auto.TypeAsync("aspire start --format json | tee /tmp/aspire-detach.json");
+        await auto.EnterAsync();
+        await auto.WaitForSuccessPromptAsync(counter, TimeSpan.FromMinutes(3));
+
+        await auto.TypeAsync("DASHBOARD_URL=$(sed -n 's/.*\"dashboardUrl\"[[:space:]]*:[[:space:]]*\"\\(https:\\/\\/localhost:[0-9]*\\).*/\\1/p' /tmp/aspire-detach.json | head -1)");
+        await auto.EnterAsync();
+        await auto.WaitForSuccessPromptAsync(counter);
+
+        await auto.TypeAsync("curl -ksSL -o /dev/null -w 'dashboard-http-%{http_code}' \"$DASHBOARD_URL\" || echo 'dashboard-http-failed'");
+        await auto.EnterAsync();
+        await auto.WaitUntilTextAsync("dashboard-http-200", timeout: TimeSpan.FromSeconds(15));
+        await auto.WaitForSuccessPromptAsync(counter);
+
+        await auto.TypeAsync("aspire stop");
+        await auto.EnterAsync();
+        await auto.WaitForSuccessPromptAsync(counter);
+
         await auto.TypeAsync("exit");
         await auto.EnterAsync();
